LayoutTests/media/modern-media-controls/media-controller/media-controller-click-to-start-expected.txt

Testing the MediaController click-to-start behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.showsStartButton is true

Dispatching click event to the start button
Media is playing
PASS mediaController.controls.showsStartButton is false

PASS successfullyParsed is true

TEST COMPLETE
